Rename profile slice to profiles to avoid confusion
