+++ src/store/rootReducer.ts
@@ -1,13 +1,13 @@
 import { combineReducers } from "redux";
 import appState from "./appState/reducer";
 import user from "./user/reducer";
-import profile from "./profile/reducer";
+import profiles from "./profiles/reducer";
 import listItems from "./listItems/reducer";
 
 export default combineReducers({
   appState,
   user,
-  profile,
+  profiles,
   listItems,
 });
